Update models.py
Set type checking on order value that's sent to the "to" method of ordered models. I found when working with this module that ordered values that are passed in as strings automatically hit the else logic. This allows for multiple model instances in an ordered set to have the same order value.