Fix for above(), below() and to() to handle certain edge cases.
In the case that an item's 'get_respect_to' values have changed
(such as an ordered item being moved from one ordered list into another), the currently
assigned order might temporarily be the same as an already existing order value.
Because of that, we cannot assume that in a case where order==self.order that nothing must be done.
One option would be to look for duplicates and if no duplicates are found, then return.
Conceptually simpler though is to just .exclude() the 'self' from the queryset
which is used for the +1/-1 field updates of all items that are before or after.

An alternative to this patch is to require the calling application to
set a temporary 'order' value (perhaps MAXINT) before calling above(),
below(), or to().