Add a management command which re-orders a OrderModel model if for some reason the ordering was corrupted.
In certain cases the models will end up in a not properly ordered state. This can be caused
by bypassing the 'delete' / 'save' methods, or when a user changes a foreign key of a object
which is part of the 'order_with_respect_to' fields.

This management command allows you to repair this invalid state.