fix: utilize order_wrt _id instead of object
the newly introduced OrderedModelBase._wrt_map method gets the
value of the fields defined in order_with_respect_to. this forces
django to query the object from the database, but only to be used
for comparison. It is optimal to just compare the _id instead of
the actual model object.