Query ForeignKey id values during load to optimise order_with_respect_to check during save()
This favours a smaller overall query count over simpler queries.

By adding attributes during evaluation of the queryset, we can store the prior value of
ForeignKey relations, for the cost of a join for each entry within
order_with_respect_to (nested joins if LOOKUP_SEPARATOR (__) is used).

The fetched attibutes are not available during OrderedModel.__init__() so we defer creation
of self._original_wrt_map until save(). Since this reads from attributes it no longer
causes any additional queries.

This seems to balance correctness in the case of modifying the fields you are ordered
with respect to, against minimising extra queries during either get() or save()
operations, and does not require the full related object to be fetched to detect the change.